fix(app): isolate keyboard shift context
